Put handle_s3_objects in its own class
